refactor: use Deno
