+++ deno.jsonc
@@ -1,0 +1,7 @@
+{
+  "compilerOptions": {
+    "checkJs": true,
+    "noUncheckedIndexedAccess": true,
+    "lib": ["dom"]
+  }
+}
