--- package.json
@@ -1,25 +1,0 @@
-{
-  "name": "svelte-app",
-  "version": "1.0.0",
-  "private": true,
-  "scripts": {
-    "build": "rollup -c",
-    "dev": "rollup -c -w",
-    "start": "sirv public --no-clear"
-  },
-  "devDependencies": {
-    "@rollup/plugin-commonjs": "22.0.2",
-    "@rollup/plugin-node-resolve": "14.1.0",
-    "rollup": "2.79.1",
-    "rollup-plugin-css-only": "3.1.0",
-    "rollup-plugin-livereload": "2.0.5",
-    "rollup-plugin-svelte": "7.1.0",
-    "rollup-plugin-terser": "7.0.2",
-    "svelte": "3.50.1"
-  },
-  "dependencies": {
-    "@guardian/libs": "8.0.5",
-    "@guardian/src-foundations": "3.13.0",
-    "sirv-cli": "2.0.2"
-  }
-}
